Add ShinyBindingWrapper to experimental.R so it can be used with :::
