+++ js/src/react/index.js
@@ -1,3 +1,4 @@
 export * from './adapters';
 export { default as findAndRenderReactData } from './findAndRenderReactData';
+export { default as ShinyBindingWrapper } from './ShinyBindingWrapper';
 export { default as ShinyProxy } from './ShinyProxy';
